+++ mathesar_ui/src/systems/record-view/DirectField.svelte
@@ -17,7 +17,7 @@
     iconExpandDown,
   } from '@mathesar-component-library';
 
-  import type RecordStore from './RecordStore';
+  import type RecordStore from '../../stores/RecordStore';
 
   /**
    * This is used to determine whether to display a `NULL` overlay indicator.
