+++ mathesar_ui/src/systems/record-view/RecordViewLoadingSpinner.svelte
