+++ mathesar_ui/src/systems/record-view/TableWidget.svelte
